Avancement sur l'analyse avec prise en charge de l'édition des liste pour la lemmatisation et la tokenization

+++ src/main/java/view/panel/ActionPanel.java
@@ -34,7 +34,7 @@ public class ActionPanel implements IActionPanel {
 	 * Permet de créer un panel action
 	 * @param nbButton nb de boutons souhaités
 	 */
-	public ActionPanel(Integer nbButton) {
+	public ActionPanel(long nbButton) {
 		for (int i = 0; i < nbButton; i++) {
 			JButton jButton = new JButton();
 			buttonMap.put(i, jButton);
